rename struct -> data_struct
