+++ src/slangpy_ext/nanobind.h
@@ -22,7 +22,7 @@
 #include "sgl/core/object.h"
 #include "sgl/core/enum.h"
 #include "sgl/core/type_utils.h"
-#include "sgl/core/struct.h"
+#include "sgl/core/data_struct.h"
 #include "sgl/core/data_type.h"
 #include "sgl/core/static_vector.h"
