+++ src/slangpy_ext/slangpy_ext.cpp
@@ -13,12 +13,12 @@ SGL_PY_DECLARE(app_app);
 
 SGL_PY_DECLARE(core_bitmap);
 SGL_PY_DECLARE(core_crypto);
+SGL_PY_DECLARE(core_data_struct);
 SGL_PY_DECLARE(core_data_type);
 SGL_PY_DECLARE(core_input);
 SGL_PY_DECLARE(core_logger);
 SGL_PY_DECLARE(core_object);
 SGL_PY_DECLARE(core_platform);
-SGL_PY_DECLARE(core_struct);
 SGL_PY_DECLARE(core_thread);
 SGL_PY_DECLARE(core_timer);
 SGL_PY_DECLARE(core_window);
@@ -97,7 +97,7 @@ NB_MODULE(slangpy_ext, m_)
     SGL_PY_IMPORT(core_logger);
     SGL_PY_IMPORT(core_timer);
     SGL_PY_IMPORT(core_window);
-    SGL_PY_IMPORT(core_struct);
+    SGL_PY_IMPORT(core_data_struct);
     SGL_PY_IMPORT(core_bitmap);
     SGL_PY_IMPORT(core_crypto);
     SGL_PY_IMPORT(core_data_type);
